Fix check for existing work dir (#1)
Currently, the work directory is always first created and then checked
for existance. This causes the update-command to not work on the first
invocation when the repositories were not yet cloned.

This PR fixes this behaviour by first checking whether the work
directory already exists and then only creating it if does not already.