Add support for surveying nested property values (#2)
This PR adds support for querying nested values in themes. Given the
context of deprecation-warnings, this might not make that much sense.
However, locally I used this to check whether some properties are
defined in themes at all. 

I also changed the loading of the JSON-files given
https://github.com/serde-rs/json/issues/160#issuecomment-253446892. For
me, this reduces the time it takes for a survey to complete from 4.5
seconds to around 450ms.